kermit-makefiles updated

+++ kermit-makefiles/scripts/bibtool/biblatex/fixbib.sh
@@ -2,7 +2,13 @@
 
 for i in `find . -name "*.bib"`
 do
-    tmpfile=`mktemp -u`
-    bibtool -i "$i" -o "$tmpfile"
-    mv "$tmpfile" "$i"
+    for rc in bibtool/*.rsc
+    do
+	if [[ -f "$rc" ]]
+	then
+	    tmpfile=`mktemp -u`
+	    bibtool -v -r "$rc" -i "$i" -o "$tmpfile"
+	    mv "$tmpfile" "$i"
+	fi
+    done
 done
